chore(mir): rename mir/emit.py → mir/print.py + test rename
Per docs/stage3a_execution_plan.md §3, mir/emit.py is Print_m (the
canonical s-expr printer for MIR), not a "Render" (which would
produce target text). The "emit" name was a holdover that conflated
Print with Render — the same conflation that S3A.0 fixed for
codegen vs dialect.

Mechanical changes:

  - git mv src/srdatalog/ir/mir/emit.py → mir/print.py
  - git mv tests/test_mir_emit_basic.py → tests/test_mir_print_basic.py
  - sed-replace srdatalog.ir.mir.emit → srdatalog.ir.mir.print
    (2 import sites: viz/bundle.py + test_mir_print_basic.py)
  - sed-replace mir/emit.py → mir/print.py in 5 docstring references
  - ruff format on resulting files

Test gate: 1057 passed, 5 skipped — baseline preserved.

This is commit 1 of Bundle A (IIR-as-data + render-registry refactor).
Subsequent commits in this PR: S3A.2 (sorted_array/lowerings returns
IIR data), S3A.3 (split codegen/cuda/emit.py into render/ package),
S3A.9b (block_group emit relocation).
