+++ src/srdatalog/viz/bundle.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from srdatalog.ir.hir.emit import hir_to_obj
-from srdatalog.ir.mir.emit import print_mir_sexpr
+from srdatalog.ir.mir.print import print_mir_sexpr
 from srdatalog.ir.pipeline import compile_program
 
 if TYPE_CHECKING:
