+++ tests/test_hir_mir_tc_e2e.py
@@ -13,7 +13,7 @@ from pathlib import Path
 
 from srdatalog.dsl import Program, Relation, Var
 from srdatalog.ir.hir import compile_to_mir
-from srdatalog.ir.mir.emit import print_mir_sexpr
+from srdatalog.ir.mir.print import print_mir_sexpr
 
 FIXTURES = Path(__file__).resolve().parent / "fixtures"
