+++ tests/test_hir_user_plan.py
@@ -12,7 +12,7 @@ from srdatalog.dsl import PlanEntry, Program, Relation, Var
 from srdatalog.ir.hir import compile_to_hir, compile_to_mir
 from srdatalog.ir.hir.emit import hir_to_obj
 from srdatalog.ir.hir.plan import derive_clause_order_from_var_order
-from srdatalog.ir.mir.emit import print_mir_sexpr
+from srdatalog.ir.mir.print import print_mir_sexpr
 
 FIXTURES = Path(__file__).resolve().parent / "fixtures"
