+++ tests/test_split_rule.py
@@ -13,7 +13,7 @@ from srdatalog.dsl import SPLIT, Program, Relation, Var
 from srdatalog.ir.hir import compile_to_hir, compile_to_mir
 from srdatalog.ir.hir.emit import hir_to_obj
 from srdatalog.ir.hir.plan import compute_temp_vars, detect_split
-from srdatalog.ir.mir.emit import print_mir_sexpr
+from srdatalog.ir.mir.print import print_mir_sexpr
 
 FIXTURES = Path(__file__).resolve().parent / "fixtures"
