+++ tests/test_split_rule_recursive.py
@@ -12,7 +12,7 @@ from pathlib import Path
 from srdatalog.dsl import SPLIT, Program, Relation, Var
 from srdatalog.ir.hir import compile_to_hir, compile_to_mir
 from srdatalog.ir.hir.emit import hir_to_obj
-from srdatalog.ir.mir.emit import print_mir_sexpr
+from srdatalog.ir.mir.print import print_mir_sexpr
 
 FIXTURES = Path(__file__).resolve().parent / "fixtures"
